move generate_state_table to test mod (test only)
